Move HomeScreenData to Data directory
